fix dependency cycle and all lint issues
